+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+*  add validation for email and phone number
+
+## 0.0.2
+*  make login type enum for login by phone number and email
+
 ## 0.0.1
 
 * TODO: Describe initial release.